Add syntax highlighting for built-in functions and Node types
- Separated Node types from built-in keywords and constants.
- Marked GDScript keywords, functions, and constants as "built-in"

I added font-faces for most of the classes in the engine, and moved classes like
Vector2 and Transform to that list, as they're classes in the engine. This is
more consistent with the way Godot's built-in script editor highlights keywords.

I also put constants like TAU and PI, although they are built-in keywords, into
their own

Moving forward, I would like to parse these directly from the engine source code
and make the tool available to everyone.